walk is very good

+++ surena4/src/trajectory_generation/src/taskspaceonline3.h
@@ -64,7 +64,7 @@ public:
     double T_beta;
 
     double Tm;
-    int Tc;
+    double Tc;
     double TDs;
     double TSS;
     double TGait;
